first attempt at a new getObject() method in the IDataSet
Benoit, if you get a chance take a look and let me know what
you think of the implementation.  I am currently getting some
NPE's in the unit tests.  This is because it is not finding
the pzconvert.properties file.  I tried to modify the ant script
to make sure it was getting included, but I am having some trouble.
I can take a look at it later and see what the deal is.

Paul

git-svn-id: https://flatpack.svn.sourceforge.net/svnroot/flatpack/trunk@174 509fa23a-e01f-0410-8d53-9f7e05c29f80